rename module names
